[Silver III] Title: 랜선 자르기, Time: 8 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -8,7 +8,7 @@ int main()
     long long st = 1, en = 0, mid;
     cin >> K >> N;
 
-    for (long long i = 0; i < K; i++)
+    for (int i = 0; i < K; i++)
     {
         cin >> line[i];
         if (en < line[i])
@@ -19,7 +19,7 @@ int main()
     {
         cnt = 0;
         mid = (st + en) / 2;
-        for (long long i = 0; i < K; i++)
+        for (int i = 0; i < K; i++)
         {
             cnt += line[i] / mid;
         }
